Move some functions into a new channel utility module
* Change `is_help_channel` to`internally use `is_in_category`
